brace work but not recursively

+++ libft/includes/mem.h
@@ -6,7 +6,7 @@
 /*   By: gavizet <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/06/02 11:21:05 by gavizet           #+#    #+#             */
-/*   Updated: 2017/11/14 17:06:16 by czalewsk         ###   ########.fr       */
+/*   Updated: 2017/12/14 17:39:17 by czalewsk         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ void				*ft_memcpy(void *dest, const void *src, size_t n);
 void				*ft_memmove(void *dest, const void *src, size_t n);
 int					ft_memcmp(const void *s1, const void *s2, size_t n);
 void				*ft_memccpy(void *dest, const void *src, int c, size_t n);
-char				**ft_realloc_tab(int size, char **tab);
+char				**ft_realloc_tab(size_t size, char **tab);
 void				*ft_memrealloc(void *addr, size_t old_size,
 		size_t new_size);
